TeukolskyQNMFunctions

Documentation for TeukolskyQNMFunctions.

TeukolskyQNMFunctions.TeukolskyQNMFunctions

TeukolskyQNMFunctions.F

TeukolskyQNMFunctions.compute_om

TeukolskyQNMFunctions.TeukolskyQNMFunctions — Module

TeukolskyQNMFunctions.jl computes the quasinormal modes and eigenfunctions for the spin s Teukolsky equation using a horizon penetrating, hyperboloidally compactified coordinate system. The main advantage of using these coordinates is that the quasinormal wavefunctions are finite valued from the black hole to future null infinity.

Currently, this code uses a Chebyshev pseudospectral method to compute the radial part of the eigenfunctions, and a spectral method to compute the angular part of the eigenfunctions The angular spectral method was originally introduced by Cook and Zalutskiy (arXiv:1410.7698).

TeukolskyQNMFunctions.F — Method

F(
  nr::myI,
  nl::myI,
  s::myI,
  lang::myI,
  mang::myI,
  a::myF,
  om::myC
)::myF

Absolute difference of Lambda seperation constant for radial and angular ODEs.

TeukolskyQNMFunctions.compute_om — Method

compute_om(
  nr::myI,
  nl::myI,
  s::myI,
  lang::myI,
  mang::myI,
  a::myF,
  om::myC; 
  tolerance::myF=tomyF(1e-6),
  epsilon::myF=tomyF(1e-6),
  gamma::myF=tomyF(1),
  verbose::Bool=false

)::Tuple{ myC, myC, Vector{myC}, Vector{myC}, Vector{myF}}

Search for quasinormal mode frequency in the complex plane using Newton's method.

Arguments

nr: number of radial Chebyshev collocation points

nl: number of spherical harmonic terms

s: spin of the field in the Teukolsky equation

lang: l angular number

mang: m angular number

a: black hole spin

om: guess for the initial quasinormal mode