TeukolskyQNMFunctions

Documentation for TeukolskyQNMFunctions.

TeukolskyQNMFunctions computes the quasinormal modes and eigenfunctions for the spin s Teukolsky equation using a horizon penetrating, hyperboloidally compactified coordinate system. The main advantage of using these coordinates is that the quasinormal wavefunctions are finite valued from the black hole to future null infinity.

Currently, this code uses a Chebyshev pseudospectral method to compute the radial part of the eigenfunctions, and a spectral method to compute the angular part of the eigenfunctions The angular spectral method was originally introduced by Cook and Zalutskiy).

TeukolskyQNMFunctions.RadialODE

TeukolskyQNMFunctions.RadialODE.Chebyshev

TeukolskyQNMFunctions.Spheroidal

TeukolskyQNMFunctions.TeukolskyQNMFunctions

TeukolskyQNMFunctions.F

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_D1

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_X

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_cheb

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_real

TeukolskyQNMFunctions.RadialODE.eig_vals_vecs

TeukolskyQNMFunctions.RadialODE.radial_discretized_eqn

TeukolskyQNMFunctions.Spheroidal.compute_A

TeukolskyQNMFunctions.Spheroidal.compute_B

TeukolskyQNMFunctions.Spheroidal.compute_C

TeukolskyQNMFunctions.Spheroidal.compute_M_matrix

TeukolskyQNMFunctions.Spheroidal.compute_l_min

TeukolskyQNMFunctions.Spheroidal.eig_vals_vecs

TeukolskyQNMFunctions.Spheroidal.mat_L

TeukolskyQNMFunctions.Spheroidal.mat_Y

TeukolskyQNMFunctions.compute_om

TeukolskyQNMFunctions.TeukolskyQNMFunctions — Module

TeukolskyQNMFunctions.jl computes the quasinormal modes and eigenfunctions for the spin s Teukolsky equation using a horizon penetrating, hyperboloidally compactified coordinate system. The main advantage of using these coordinates is that the quasinormal wavefunctions are finite valued from the black hole to future null infinity.

TeukolskyQNMFunctions.F — Method

function F(
    nr::Integer,
    nl::Integer,
    s::Integer,
    l::Integer,
    m::Integer,
    a::T,
    om::Complex{T},
) where {T<:Real}

Compute the absolute difference of Lambda seperation constant for radial and angular ODEs.

TeukolskyQNMFunctions.compute_om — Method

compute_om(
    nr::Integer,
    nl::Integer,
    s::Integer,
    l::Integer,
    m::Integer,
    a::Real,
    om::Complex;
    tolerance::Real = 1e6,
    epsilon::Real = 1e-6,
    gamma::Real = 1.0,
    verbose::Bool = false
) where T<:Real

Search for quasinormal mode frequency in the complex plane using Newton's method.

Arguments

nr : number of radial Chebyshev collocation points

nl : number of spherical harmonic terms

s : spin of the field in the Teukolsky equation

l : l angular number

m : m angular number

a : black hole spin

om : guess for the initial quasinormal mode

tolerance: tolerance for root finder

epsilon : derivative finite difference length

gamma : search gamma

verbose : true: print out intermediate results as searches for root

Modules = [TeukolskyQNMFunctions.CustomTypes]

TeukolskyQNMFunctions.Spheroidal — Module

Methods to compute spin-weighted spheroidal harmonics

TeukolskyQNMFunctions.Spheroidal.compute_A — Method

compute_A(s::Integer, l::Integer, m::Integer)

See Appendix A of J. Ripley, Class.Quant.Grav. 39 (2022) 14, 145009, Class.Quant.Grav. 39 (2022) 145009 arXiv:2202.03837

TeukolskyQNMFunctions.Spheroidal.compute_B — Method

compute_B(s::Integer, l::Integer, m::Integer)

See Appendix A of J. Ripley, Class.Quant.Grav. 39 (2022) 14, 145009, Class.Quant.Grav. 39 (2022) 145009 arXiv:2202.03837

TeukolskyQNMFunctions.Spheroidal.compute_C — Method

compute_C(s::Integer, l::Integer, m::Integer)

See Appendix A of J. Ripley, Class.Quant.Grav. 39 (2022) 14, 145009, Class.Quant.Grav. 39 (2022) 145009 arXiv:2202.03837

TeukolskyQNMFunctions.Spheroidal.compute_M_matrix — Method

compute_M_matrix(nl::Integer, s::Integer, m::Integer, c::Complex)

Compute the matrix for computing spheroidal-spherical mixing and seperation coefficients.

TeukolskyQNMFunctions.Spheroidal.compute_l_min — Method

compute_l_min(s::Integer, m::Integer)

Compute the minimum l value

TeukolskyQNMFunctions.Spheroidal.eig_vals_vecs — Method

eig_vals_vecs(nl::Integer, neig::Integer, s::Integer, m::Integer, c::Complex)

Compute the eigenvectors and eigenvalues for the spheroidal equation. Returns eigenvalues λs (smallest to largest), and eigenvectors as an array (access nth eigenvector through v[:,n]); i.e. returns (λs, vs)

TeukolskyQNMFunctions.Spheroidal.mat_L — Method

function mat_L(nl::Integer, s::Integer, m::Integer)

Compute the matrix for spherical laplacian (s=0) in spectral space.

TeukolskyQNMFunctions.Spheroidal.mat_Y — Method

mat_Y(nl::Integer, s::Integer, m::Integer)

Compute the matrix for y in spectral space.

TeukolskyQNMFunctions.RadialODE — Module

Ordinary differential equation in radial direction for the hyperboloidally compactified Teukolsky equation.

TeukolskyQNMFunctions.RadialODE.eig_vals_vecs — Method

eig_vals_vecs(
        nr::Integer,
        s::Integer,
        m::Integer,
        a::T,
        om::Complex{T}
    ) where T<:Real

Compute eigenvectors and eigenvalues for the radial equation using a pseudospectral Chebyshev polynomial method. The black hole mass is always one.

TeukolskyQNMFunctions.RadialODE.radial_discretized_eqn — Method

function radial_discretized_eqn(
    nr::Integer,
    s::Integer,
    m::Integer,
    a::T,
    bhm::T,
    om::Complex{T},
    rmin::T,
    rmax::T
) where T<:Real

TeukolskyQNMFunctions.RadialODE.Chebyshev — Module

Methods for position space Chebyshev methods.

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts — Method

cheb_pts(xmin::Real, xmax::Real, nx::Integer)

Computes Chebyshev points on interval [xmin,xmax]

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts — Method

cheb_pts([T,] nx::Integer)

Computes Generates nx Chebyshev points in [-1,1]

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_D1 — Method

mat_D1(
      xmin::Real,
      xmax::Real,
      nx::Integer,
   )

Computes derivative matrix D1 in real space.

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_X — Method

mat_X(
      xmin::Real,
      xmax::Real,
      nx::Integer
   )

Computes matrix for multiplication of x in real space.

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_cheb — Method

to_cheb(f::Vector{<:Number})

Convert to Chebyshev space. We assume we are working with Chebyshev-Gauss-Lobatto points.

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_real — Method

to_real(c::Vector{<:Number})

Convert to Real space. We assume we are working with Chebyshev-Gauss-Lobatto points.