TeukolskyQNMFunctions

Documentation for TeukolskyQNMFunctions.

TeukolskyQNMFunctions computes the quasinormal modes and eigenfunctions for the spin s Teukolsky equation using a horizon penetrating, hyperboloidally compactified coordinate system. The main advantage of using these coordinates is that the quasinormal wavefunctions are finite valued from the black hole to future null infinity.

Currently, this code uses a Chebyshev pseudospectral method to compute the radial part of the eigenfunctions, and a spectral method to compute the angular part of the eigenfunctions The angular spectral method was originally introduced by Cook and Zalutskiy).

TeukolskyQNMFunctions.CustomTypes

TeukolskyQNMFunctions.RadialODE

TeukolskyQNMFunctions.RadialODE.Chebyshev

TeukolskyQNMFunctions.Spheroidal

TeukolskyQNMFunctions.TeukolskyQNMFunctions

TeukolskyQNMFunctions.F

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_D1

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_X

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_cheb

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_real

TeukolskyQNMFunctions.RadialODE.eig_vals_vecs

TeukolskyQNMFunctions.Spheroidal.compute_A

TeukolskyQNMFunctions.Spheroidal.compute_B

TeukolskyQNMFunctions.Spheroidal.compute_C

TeukolskyQNMFunctions.Spheroidal.compute_M_matrix

TeukolskyQNMFunctions.Spheroidal.compute_l_min

TeukolskyQNMFunctions.Spheroidal.eig_vals_vecs

TeukolskyQNMFunctions.Spheroidal.mat_L

TeukolskyQNMFunctions.Spheroidal.mat_Y

TeukolskyQNMFunctions.compute_om

TeukolskyQNMFunctions.TeukolskyQNMFunctions — Module

TeukolskyQNMFunctions.jl computes the quasinormal modes and eigenfunctions for the spin s Teukolsky equation using a horizon penetrating, hyperboloidally compactified coordinate system. The main advantage of using these coordinates is that the quasinormal wavefunctions are finite valued from the black hole to future null infinity.

TeukolskyQNMFunctions.F — Method

F(
  nr::myI,
  nl::myI,
  s::myI,
  lang::myI,
  mang::myI,
  a::myF,
  om::myC
)::myF

Absolute difference of Lambda seperation constant for radial and angular ODEs.

TeukolskyQNMFunctions.compute_om — Method

compute_om(
  nr::myI,
  nl::myI,
  s::myI,
  lang::myI,
  mang::myI,
  a::myF,
  om::myC; 
  tolerance::myF=tomyF(1e-6),
  epsilon::myF=tomyF(1e-6),
  gamma::myF=tomyF(1),
  verbose::Bool=false

)::Tuple{ myC, myC, Vector{myC}, Vector{myC}, Vector{myF}}

Search for quasinormal mode frequency in the complex plane using Newton's method.

Arguments

nr: number of radial Chebyshev collocation points

nl: number of spherical harmonic terms

s: spin of the field in the Teukolsky equation

lang: l angular number

mang: m angular number

a: black hole spin

om: guess for the initial quasinormal mode

TeukolskyQNMFunctions.CustomTypes — Module

Custom number types for extended-precision computations.

TeukolskyQNMFunctions.Spheroidal — Module

Methods to compute spin-weighted spheroidal harmonics

TeukolskyQNMFunctions.Spheroidal.compute_A — Method

compute_A(s::myI, l::myI, m::myI)

See the note.

TeukolskyQNMFunctions.Spheroidal.compute_B — Method

compute_B(s::myI, l::myI, m::myI)

See the note.

TeukolskyQNMFunctions.Spheroidal.compute_C — Method

compute_C(s::myI, l::myI, m::myI)

See the note.

TeukolskyQNMFunctions.Spheroidal.compute_M_matrix — Method

compute_M_matrix(
  nl::myI,
  s::myI,
  m::myI,
  c::myC
)

M matrix for computing spheroidal-spherical mixing and seperation coefficients.

TeukolskyQNMFunctions.Spheroidal.compute_l_min — Method

compute_l_min(s,m)

Minimum l value

TeukolskyQNMFunctions.Spheroidal.eig_vals_vecs — Method

eig_vals_vecs(
  nl::myI,
  neig::myI,
  s::myI,
  m::myI,
  c::myC
)

Compute eigenvectors and eigenvalues for the spheroidal equation. Returns eigenvalues λs (smallest to largest), and eigenvectors as an array (access nth eigenvector through v[:,n]); i.e. returns (λs, vs)

TeukolskyQNMFunctions.Spheroidal.mat_L — Method

mat_L(
  nl::myI,
  s::myI,
  m::myI,
)::SparseMatrixCSC

Returns the matrix for spherical laplacian (s=0) in spectral space.

TeukolskyQNMFunctions.Spheroidal.mat_Y — Method

mat_Y(
  nl::myI,
  s::myI,
  m::myI,
)::SparseMatrixCSC

Returns the matrix for y in spectral space.

TeukolskyQNMFunctions.RadialODE — Module

Ordinary differential equation in radial direction for the hyperboloidally compactified Teukolsky equation.

TeukolskyQNMFunctions.RadialODE.eig_vals_vecs — Method

eig_vals_vecs(
  nr::myI,
  s::myI,
  mang::myI,
  a::myF,
  om::myC
)

Compute eigenvectors and eigenvalues for the radial equation. The black hole mass is always one.

TeukolskyQNMFunctions.RadialODE.Chebyshev — Module

Methods for position space Chebyshev methods.

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts — Method

cheb_pts(xmin::myF, xmax::myF, nx::myI)::Vector{myF}

Computes Chebyshev points on interval [xmin,xmax]

TeukolskyQNMFunctions.RadialODE.Chebyshev.cheb_pts — Method

cheb_pts(nx::myI)::Vector{myF}

Computes Generates nx Chebyshev points in [-1,1]

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_D1 — Method

mat_D1(
  xmin::myF,
  xmax::myF,
  nx::myI
)::Matrix{myF}

Computes derivative matrix D1 in real space.

TeukolskyQNMFunctions.RadialODE.Chebyshev.mat_X — Method

mat_X(
  xmin::myF,
  xmax::myF,
  nx::myI
)::SparseMatrixCSC{myF, myI}

Computes matrix for multiplication of x in real space.

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_cheb — Method

to_cheb(f::Vector{T})::Vector{T} where T<:Number

Convert to Chebyshev space. We assume we are working with Chebyshev-Gauss-Lobatto points.

TeukolskyQNMFunctions.RadialODE.Chebyshev.to_real — Method

to_real(c::Vector{T})::Vector{T} where T<:Number

Convert to Real space. We assume we are working with Chebyshev-Gauss-Lobatto points.